Starting FMS/LNAV refactor
