+++ sauna-api/ApiObjects/Aircraft/AircraftResponse.cs
@@ -2,8 +2,6 @@ using SaunaSim.Core;
 using SaunaSim.Core.Data;
 using SaunaSim.Core.Simulator.Aircraft;
 using SaunaSim.Core.Simulator.Aircraft.Control;
-using SaunaSim.Core.Simulator.Aircraft.Control.FMS;
-using SaunaSim.Core.Simulator.Aircraft.Control.FMS.Legs;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 using System.Collections.Generic;
@@ -22,10 +20,8 @@ namespace SaunaSim.Api.ApiObjects.Aircraft
         public TransponderModeType XpdrMode { get; set; }
         public int Squawk { get; set; }
         public AircraftPosition Position { get; set; }
-        public AircraftControlResponse Control { get; set; }
+        public AircraftFmsResponse Fms { get; set; }
         public AircraftAutopilot Autopilot { get; set; }
-        public int Assigned_IAS { get; set; }
-        public ConstraintType Assigned_IAS_Type { get; set; }
         public ConnectionStatusType ConnectionStatus { get; set; }
 
         public AircraftResponse()
@@ -42,51 +38,13 @@ namespace SaunaSim.Api.ApiObjects.Aircraft
             XpdrMode = pilot.XpdrMode;
             Squawk = pilot.Squawk;
             Position = pilot.Position;
-            Control = new AircraftControlResponse(pilot.Control, includeFms);
-            Assigned_IAS = pilot.Assigned_IAS;
-            Assigned_IAS_Type = pilot.Assigned_IAS_Type;
+            if (includeFms)
+            {
+                Fms = new AircraftFmsResponse(pilot.Fms);
+            }
             ConnectionStatus = pilot.ConnectionStatus;
             ThrustLeverPos = pilot.ThrustLeverPos;
             Autopilot = pilot.Autopilot;
-        }
-
-        public class AircraftControlResponse
-        {
-            public object CurrentLateralMode { get; set; }
-            public string CurrentLateralModeStr { get; set; }
-            public object ArmedLateralMode { get; set; }
-            public string ArmedLateralModeStr { get; set; }
-            public object CurrentVerticalMode { get; set; }
-            public string CurrentVerticalModeStr { get; set; }
-            public List<object> ArmedVerticalModes { get; set; }
-            public List<string> ArmedVerticalModesStr { get; set; }
-            public AircraftFmsResponse FMS { get; set; }
-
-            public AircraftControlResponse()
-            {
-
-            }
-
-            public AircraftControlResponse(AircraftControl control, bool includeFms = false)
-            {
-                CurrentLateralMode = control.CurrentLateralInstruction;
-                CurrentLateralModeStr = control.CurrentLateralInstruction?.ToString();
-                ArmedLateralMode = control.ArmedLateralInstruction;
-                ArmedLateralModeStr = control.ArmedLateralInstruction?.ToString();
-                CurrentVerticalMode = control.CurrentVerticalInstruction;
-                CurrentVerticalModeStr = control.CurrentVerticalInstruction?.ToString();
-                ArmedVerticalModes = new List<object>();
-                ArmedVerticalModesStr = new List<string>();
-                foreach (var instr in control.ArmedVerticalInstructions)
-                {
-                    ArmedVerticalModes.Add(instr);
-                    ArmedVerticalModesStr.Add(instr.ToString());
-                }
-                if (includeFms)
-                {
-                    FMS = new AircraftFmsResponse(control.FMS);
-                }
-            }
         }
     }
 }
