+++ sauna-api/ApiObjects/Aircraft/FmsWaypointRequest.cs
@@ -1,7 +1,7 @@
 
 using SaunaSim.Core;
 using SaunaSim.Core.Simulator.Aircraft;
-using SaunaSim.Core.Simulator.Aircraft.Control.FMS;
+using SaunaSim.Core.Simulator.Aircraft.FMS;
 
 namespace SaunaSim.Api.ApiObjects.Aircraft
 {
