+++ sauna-sim-core/Data/Procedures/Approaches/ApproachProcedure.cs
@@ -5,7 +5,6 @@ using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
 using SaunaSim.Core.Simulator.Aircraft;
-using SaunaSim.Core.Simulator.Aircraft.Control.FMS;
 
 namespace SaunaSim.Core.Data.Procedures.Approaches
 {
