--- sauna-sim-core/Simulator/Aircraft/Control/FMS/IRoutePoint.cs
@@ -1,16 +1,0 @@
-﻿using AviationCalcUtilNet.GeoTools;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SaunaSim.Core.Simulator.Aircraft.Control.FMS
-{
-    public interface IRoutePoint
-    {
-        GeoPoint PointPosition { get; }
-
-        string PointName { get; }
-    }
-}
