+++ sauna-sim-core/Simulator/Aircraft/Control/Instructions/Lateral/IlsApproachInstruction.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using SaunaSim.Core.Data;
-using SaunaSim.Core.Simulator.Aircraft.Control.FMS;
+using SaunaSim.Core.Simulator.Aircraft.FMS;
 
 namespace SaunaSim.Core.Simulator.Aircraft.Control.Instructions.Lateral
 {
