+++ sauna-sim-core/Simulator/Aircraft/Control/Instructions/Lateral/TrackHoldInstruction.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using SaunaSim.Core.Simulator.Aircraft.Control;
-using SaunaSim.Core.Simulator.Aircraft.Control.FMS;
+using SaunaSim.Core.Simulator.Aircraft.FMS;
 
 namespace SaunaSim.Core.Simulator.Aircraft
 {
