+++ sauna-sim-core/Simulator/Aircraft/FMS/AircraftFms.cs
@@ -1,12 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using SaunaSim.Core.Data;
-using SaunaSim.Core.Simulator.Aircraft.Control.FMS.Legs;
+using SaunaSim.Core.Simulator.Aircraft.FMS.Legs;
 
-namespace SaunaSim.Core.Simulator.Aircraft.Control.FMS
+namespace SaunaSim.Core.Simulator.Aircraft.FMS
 {
     public class AircraftFms
     {
