+++ sauna-sim-core/Simulator/Aircraft/FMS/FmsPoint.cs
@@ -1,11 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using SaunaSim.Core.Data;
-
-namespace SaunaSim.Core.Simulator.Aircraft.Control.FMS
+﻿namespace SaunaSim.Core.Simulator.Aircraft.FMS
 {
     public enum RoutePointTypeEnum
     {
