+++ sauna-sim-core/Simulator/Aircraft/FMS/IRoutePoint.cs
@@ -1,0 +1,11 @@
+﻿using AviationCalcUtilNet.GeoTools;
+
+namespace SaunaSim.Core.Simulator.Aircraft.FMS
+{
+    public interface IRoutePoint
+    {
+        GeoPoint PointPosition { get; }
+
+        string PointName { get; }
+    }
+}
