+++ sauna-sim-core/Simulator/Aircraft/FMS/Legs/FixToManualLeg.cs
@@ -1,25 +1,17 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace SaunaSim.Core.Simulator.Aircraft.Control.FMS.Legs
+namespace SaunaSim.Core.Simulator.Aircraft.FMS.Legs
 {
-    public class FixToAltLeg : IRouteLeg
+    public class FixToManualLeg : IRouteLeg
     {
         private FmsPoint _startPoint;
         private double _magneticCourse;
         private double _trueCourse;
-        private double _endAlt;
-        private double _beginAlt;
         private InterceptCourseInstruction _instr;
 
-        public FixToAltLeg(FmsPoint startPoint, BearingTypeEnum courseType, double course, double endAlt)
+        public FixToManualLeg(FmsPoint startPoint, BearingTypeEnum courseType, double course)
         {
             _startPoint = startPoint;
-            _endAlt = endAlt;
-            _beginAlt = -1;
 
             if (courseType == BearingTypeEnum.TRUE)
             {
@@ -46,20 +38,12 @@ namespace SaunaSim.Core.Simulator.Aircraft.Control.FMS.Legs
 
         public ILateralControlInstruction Instruction => _instr;
 
-        public RouteLegTypeEnum LegType => RouteLegTypeEnum.FIX_TO_ALT;
+        public RouteLegTypeEnum LegType => RouteLegTypeEnum.FIX_TO_MANUAL;
 
         public bool HasLegTerminated(AircraftPosition pos, ref AircraftFms fms)
         {
-            if (_beginAlt < 0)
-            {
-                _beginAlt = pos.IndicatedAltitude;
-            }
-
-            if (_beginAlt <= _endAlt)
-            {
-                return pos.IndicatedAltitude >= _endAlt;
-            }
-            return pos.IndicatedAltitude <= _endAlt;
+            // Manual termination
+            return false;
         }
 
         public bool ShouldBeginTurn(AircraftPosition pos, AircraftFms fms, int posCalcIntvl)
@@ -69,24 +53,7 @@ namespace SaunaSim.Core.Simulator.Aircraft.Control.FMS.Legs
 
         public void UpdateLateralPosition(ref AircraftPosition pos, ref AircraftFms fms, int posCalcIntvl)
         {
-            if (_beginAlt < 0)
-            {
-                _beginAlt = pos.IndicatedAltitude;
-            }
-
-            IRouteLeg nextLeg = fms.GetFirstLeg();
-
-            // Only sequence if next leg exists and fms is not suspended
-            if (nextLeg != null && !fms.Suspended)
-            {
-                if (HasLegTerminated(pos, ref fms))
-                {
-                    // Activate next leg on termination
-                    fms.ActivateNextLeg();
-                }
-            }
-
-            // Otherwise update position as normal
+            // Will not auto sequence
             _instr.UpdatePosition(ref pos, ref fms, posCalcIntvl);
         }
 
@@ -97,7 +64,7 @@ namespace SaunaSim.Core.Simulator.Aircraft.Control.FMS.Legs
 
         public override string ToString()
         {
-            return $"{_startPoint.Point.PointName}-{_magneticCourse:000} =(FA)=> {_endAlt}";
+            return $"{_startPoint.Point.PointName}-{_magneticCourse:000} =(FM)=> MANUAL";
         }
     }
 }
